feat(refactor): renamed file names and fixed imports accordingly and changed some components to FC
