+++ src/components/contributorpage/NewContributrs.tsx
@@ -1,5 +1,5 @@
 import { ContributorsProps } from '@/pages/contributors';
-import FresLeafIcon from '@icons/fresh-leaf';
+import FresLeafIcon from '@icons/FreshLeaf';
 import { FC } from 'react';
 import Represent from '../layouts/represent';
 import RenderContributors from './RenderContributors';
