+++ src/components/contributorpage/OldContributrs.tsx
@@ -1,5 +1,5 @@
 import { ContributorsProps } from '@/pages/contributors';
-import EarlyBirdsIcon from '@icons/early-birds';
+import EarlyBirdsIcon from '@icons/EarlyBirds';
 import { FC } from 'react';
 import Represent from '../layouts/represent';
 import RenderContributors from './RenderContributors';
