+++ src/components/footer/FooterList.tsx
@@ -1,5 +1,5 @@
 import HeroLinks from '@/data/HeroLinks';
-import { links } from '@/data/menu-links';
+import { links } from '@/data/MenuLinks';
 import Link from 'next/link';
 import { FC } from 'react';
