+++ src/components/hero/Links.tsx
@@ -1,4 +1,5 @@
 import HeroLinks from '@/data/HeroLinks';
+import { FC } from 'react';
 const replaceData = ['Create a Copy', 'Go to Repository'];
 
 const HLinks = HeroLinks.map((link, indx) => {
@@ -8,7 +9,7 @@ const HLinks = HeroLinks.map((link, indx) => {
   };
 });
 
-export default function Links() {
+const Links: FC = () => {
   return (
     <div className="flex flex-wrap-reverse items-center gap-y-4 gap-x-4 font-semibold text-skin-base sm:gap-x-6">
       {HLinks.map((link) => (
@@ -25,4 +26,5 @@ export default function Links() {
       ))}
     </div>
   );
-}
+};
+export default Links;
