+++ src/components/icons/EarlyBirds.tsx
