+++ src/components/icons/StarMan.tsx
