+++ src/pages/contributors/index.tsx
@@ -2,8 +2,8 @@ import type { TCont } from '&validation/contributor.validation';
 import NewContributrs from '@/components/contributorpage/NewContributrs';
 import OldContributrs from '@/components/contributorpage/OldContributrs';
 import PopularContributrs from '@/components/contributorpage/PopularContributrs';
-import Container from '@layouts/container';
-import Layout from '@layouts/main';
+import Container from '@layouts/Container';
+import Layout from '@layouts/Main';
 import {
   fetchNewContributors,
   fetchOldContributors,
